Create 3.8 Leetcode question 3
This C++ program calculates the bitwise complement of a given integer. It handles the edge case when n is 0 by returning 1, then uses bit manipulation to construct a mask that matches the bit length of n. The final result is obtained by applying the bitwise NOT to n and ANDing it with the mask. Clean use of bit shifts and masking to limit the complement to the relevant bits.